# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for learning English via dictation (video-based).",
        "title": "Dictation Learning System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/v1",
    "paths": {
...
        "/lessons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay danh sach bai hoc co phan trang; ho tro loc theo do kho va category.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Danh sach bai hoc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by level (easy|medium|hard)",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay danh sach bai hoc thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_pkg_utils.ListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tao bai hoc moi tu payload frontend (url + transcripts), va co the gan category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Tao bai hoc",
                "parameters": [
                    {
                        "description": "Du lieu bai hoc tu frontend",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-familytree_internal_service.CreateLessonInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tao bai hoc thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_models.Lesson"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/lessons/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay thong tin chi tiet cua mot bai hoc theo ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Chi tiet bai hoc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay chi tiet bai hoc thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_models.Lesson"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/lessons/{id}/transcripts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay toan bo transcript cua bai hoc, sap xep theo thu tu sequence.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Danh sach transcript cua bai hoc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay transcript thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/go-familytree_internal_models.Transcript"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-familytree_internal_models.Lesson": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "description": "easy|medium|hard",
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "transcripts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-familytree_internal_models.Transcript"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "go-familytree_internal_models.Transcript": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "end_timestamp": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "phonetic": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "start_timestamp": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "vietnamese": {
                    "type": "string"
                }
            }
        },
...
        "go-familytree_internal_service.CreateLessonInput": {
            "type": "object",
            "required": [
                "title",
                "transcripts"
            ],
            "properties": {
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "number",
                    "minimum": 0
                },
                "level": {
                    "type": "string",
                    "enum": [
                        "easy",
                        "medium",
                        "hard"
                    ]
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "transcripts": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/go-familytree_internal_service.CreateTranscriptInput"
                    }
                },
                "url": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "go-familytree_internal_service.CreateTranscriptInput": {
            "type": "object",
            "required": [
                "content",
                "end_timestamp",
                "sequence",
                "start_timestamp"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "end_timestamp": {
                    "type": "number",
                    "minimum": 0
                },
                "phonetic": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer",
                    "minimum": 1
                },
                "start_timestamp": {
                    "type": "number",
                    "minimum": 0
                },
                "vietnamese": {
                    "type": "string"
                }
            }
        },
...
        "go-familytree_pkg_response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "go-familytree_pkg_utils.ListResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "meta": {
                    "$ref": "#/definitions/go-familytree_pkg_utils.PaginationMeta"
                }
            }
        },
        "go-familytree_pkg_utils.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}